Mobile app bug fixes
Bug 1 - dont_send_user_credentials was not being pulled from default settings when creating new mobile app organization

Bug 2 - #56 - when erasing an organization, the extension UUID data associated with it was not being deleted. As a result, if mobile apps were to be added to that same organization at a later time, "no organization found" would appear when trying to enable mobile apps for extensions that previously had it connected

+++ app/Http/Controllers/AppsController.php
@@ -584,6 +584,10 @@ class AppsController extends Controller
             $deleteResponse = $this->ringotelApiService->deleteOrganization($org_id);
 
             if ($deleteResponse) {
+                MobileAppUsers::where('domain_uuid', $domain_uuid)
+                    ->where('org_id', $org_id)
+                    ->delete();
+
                 return response()->json([
                     'messages' => ['success' => ['Organization and its connections were successfully deleted.']]
                 ], 200); // 200 OK
